Begin friends list code
