+++ src/scripts/dashboard.js
@@ -1,5 +1,6 @@
 import API from "./data.js"
 import renderToDom from "./render.js"
+import createMessageObject from "./createMessage.js"
 
 /* File created/refactored by David Larsen
     Original "load dashboard" code by Brett Stoudt
@@ -12,11 +13,62 @@ const loadDashboard = () => {
     registrationContainer.classList.toggle("hidden")
     hiddenDashboard.classList.toggle("hidden")
     
+    // Queue current user ID
+    const currentUserID = parseInt(sessionStorage.getItem("currentUser"))
     // Load individual data components
+    // API.getFriendsData(currentUserID).then(friendsArray => renderToDom.friendsList(friendsArray))
     API.getMessagesData().then(messagesCollection => renderToDom.messagesList(messagesCollection))
+    .then(() => {
+        // Save new message
+
+        document.querySelector(".messagesSubmitButton").addEventListener("click", event => {
+            let messageField = document.getElementById("messagesUserInput").value
+            let messageId = document.getElementById("messageId").value
+            const messageToSave = createMessageObject(messageField)
+
+            // Check if message is new or edited
+            if (messageId === "") {
+
+                API.newMessagesEntry(messageToSave)
+                    .then(() => API.getMessagesData()
+                        .then(messagesCollection => {
+                            // Clear the message blank and refresh messages list
+                            messageField = ""
+                            renderToDom.messagesList(messagesCollection)
+                        }))
+
+            } else {
+                // Edited entry
+                API.editMessage(messageToSave, messageId)
+                    .then(() => API.getMessagesData()
+                        .then(messagesCollection => {
+                            // Clear the message blank and refresh messages list
+                            messageField = ""
+                            messageId = ""
+                            renderToDom.messagesList(messagesCollection)
+                        })
+                    )
+            }
+
+        })
+
+        // Edit existing message
+        document.querySelector(".messagesList").addEventListener("click", event => {
+            if (event.target.id.startsWith("editMessage--")) {
+                const entryIdToEdit = event.target.id.split("--")[1]
+                fetch(`http://localhost:8088/messages/${entryIdToEdit}`)
+                    .then(response => response.json())
+                    .then(entry => {
+                        document.getElementById("messagesUserInput").value = entry.message
+                        document.getElementById("messageId").value = entry.id
+                    })
+            }
+        })
+    })
     renderToDom.tasksList()
     renderToDom.eventsList()
     renderToDom.articlesList()
+
 
 }
