+++ src/scripts/render.js
@@ -1,6 +1,5 @@
 import API from "./data.js"
 import domObject from "./domobject.js"
-import createMessageObject from "./createMessage.js"
 
 /* Object which renders data to each section of the Dashboard
 
@@ -18,53 +17,6 @@ const renderToDom = {
         messagesArray.forEach(message => {
             const messageHTML = domObject.messageComponent(message)
             messagesListDOM.innerHTML += messageHTML
-        })
-
-        // Save new message
-
-        document.querySelector(".messagesSubmitButton").addEventListener("click", event => {
-            let messageField = document.getElementById("messagesUserInput").value
-            let messageId = document.getElementById("messageId").value
-            const messageToSave = createMessageObject(messageField)
-
-            // Check if message is new or edited
-            if (messageId === "") {
-
-                API.newMessagesEntry(messageToSave)
-                .then(() => API.getMessagesData()
-                .then(messagesCollection => {
-                    // Clear the message blank and refresh messages list
-                    messageField = ""
-                    renderToDom.messagesList(messagesCollection)
-                }))
-
-            } else {
-                // Edited entry
-                API.editMessage(messageToSave, messageId)
-                    .then(() => API.getMessagesData()
-                        .then(messagesCollection => {
-                            // Clear the message blank and refresh messages list
-                            messageField = ""
-                            messageId = ""
-                            renderToDom.messagesList(messagesCollection)
-                        }
-                    )
-                )
-            }
-                
-        })
-
-        // Edit existing message
-        messagesListDOM.addEventListener("click", event => {
-            if (event.target.id.startsWith("editMessage--")) {
-                const entryIdToEdit = event.target.id.split("--")[1]
-                fetch(`http://localhost:8088/messages/${entryIdToEdit}`)
-                    .then(response => response.json())
-                    .then(entry => {
-                        document.getElementById("messagesUserInput").value = entry.message
-                        document.getElementById("messageId").value = entry.id
-                    })
-            }
         })
 
     },
@@ -168,7 +120,11 @@ const renderToDom = {
     },
 
     friendsList(friendsArray) {
-
+        friendsArray.forEach(friend => {
+            const friendHTML = domObject.friendComponent(friend)
+            document.querySelector(".friendsList").innerHTML += friendHTML
+        })
+  
     }
 }
